cogitatio-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cogitatio"
version = "0.1.0"
description = "Cogitatio Virtualis Backend Server"
requires-python = ">=3.8"
dependencies = [
    "fastapi~=0.104.1",           # Updated to latest stable
    "pydantic~=2.5.1",            # Updated to match FastAPI requirements
    "uvicorn[standard]~=0.24.0",  # Updated
    "python-dotenv>=0.19.0",
    "voyageai>=0.1.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.21.0",
    "python-multipart>=0.0.5",
    "SQLAlchemy>=1.4.23",
    "PyYAML>=5.4.1",
    "watchdog>=2.1.6",
    "tabulate>=0.8.9",
    "scikit-learn>=0.24.2",
    "Flask>=2.0.1",
    "plotly>=5.1.0"
]

[project.scripts]
cogitatio-server = "scripts.start_server:main"

[tool.hatch.build.targets.wheel]
packages = ["cogitatio"]